Rename to <foo>Middleware
Otherwise it can be very annoying to use, as `cache`, `path`, `json`
etc. are also common package and/or variable names.

It's a bit ugly and mouthful, but these packages are usually only used
once or twice, so it's probably okay (we have the same pattern with
controlers in desk: `ticketsContoller`, `customerController`, etc.)