Fix: Rate limit algorithm not covering the last request from the rate
Change the remaining to also consider zero as granted to cover the last request
from the rate.

The current algorithm do the following:
1. Add entry (from the current request) to Redis
2. Remove old entries (outside the period)
3. Count the number of entries

If we allow 1 request per minute, and send a request every 60 seconds, as we
first add it to Redis, the step 3 would return 1, and the calculation from
`perPeriod - len(keys)` will return 0. The current implementation considers
zero as not granted when it should.